+++ fisica/Emec.js
@@ -1,6 +1,6 @@
 function emecbasica() {
-  let Ec = parseInt(prompt("Ingresa la Energía Cinética"));
-  let Ep = parseInt(prompt("Intresa la Energía Mecánica"));
+  let Ec = parseFloat(prompt("Ingresa la Energía Cinética"));
+  let Ep = parseFloat(prompt("Intresa la Energía Mecánica"));
   let r = parseInt(Ec) + parseInt(Ep);
   swal(
     `ENERGIA MECÁNICA`,
